+++ src/driver/sqlite/SqliteQueryRunner.ts
@@ -13,6 +13,8 @@ import {RandomGenerator} from "../../util/RandomGenerator";
 import {SqliteDriver} from "./SqliteDriver";
 import {Connection} from "../../connection/Connection";
 import {ReadStream} from "fs";
+import {OrmUtils} from "../../util/OrmUtils";
+import {InsertResult} from "../InsertResult";
 
 /**
  * Runs queries on a single sqlite database connection.
@@ -158,14 +160,15 @@ export class SqliteQueryRunner implements QueryRunner {
      * Insert a new row with given values into the given table.
      * Returns value of the generated column if given and generate column exist in the table.
      */
-    async insert(tableName: string, keyValues: ObjectLiteral, generatedColumn?: ColumnMetadata): Promise<any> {
+    async insert(tableName: string, keyValues: ObjectLiteral): Promise<InsertResult> {
         const keys = Object.keys(keyValues);
         const columns = keys.map(key => `"${key}"`).join(", ");
         const values = keys.map((key, index) => "$" + (index + 1)).join(",");
+        const generatedColumns = this.connection.hasMetadata(tableName) ? this.connection.getMetadata(tableName).generatedColumns : [];
         const sql = columns.length > 0 ? (`INSERT INTO "${tableName}"(${columns}) VALUES (${values})`) : `INSERT INTO "${tableName}" DEFAULT VALUES`;
         const parameters = keys.map(key => keyValues[key]);
 
-        return new Promise<any[]>(async (ok, fail) => {
+        return new Promise<InsertResult>(async (ok, fail) => {
             this.driver.connection.logger.logQuery(sql, parameters, this);
             const __this = this;
             const databaseConnection = await this.connect();
@@ -175,10 +178,16 @@ export class SqliteQueryRunner implements QueryRunner {
                     __this.driver.connection.logger.logQueryError(err, this);
                     fail(err);
                 } else {
-                    if (generatedColumn)
-                        return ok(this["lastID"]);
+                    const generatedMap = generatedColumns.reduce((map, generatedColumn) => {
+                        const value = generatedColumn.isPrimary && generatedColumn.generationStrategy === "increment" && this["lastID"] ? this["lastID"] : keyValues[generatedColumn.databaseName];
+                        if (!value) return map;
+                        return OrmUtils.mergeDeep(map, generatedColumn.createValueMap(value));
+                    }, {} as ObjectLiteral);
 
-                    ok();
+                    ok({
+                        result: undefined,
+                        generatedMap: Object.keys(generatedMap).length > 0 ? generatedMap : undefined
+                    });
                 }
             });
         });
@@ -651,7 +660,7 @@ export class SqliteQueryRunner implements QueryRunner {
             c += " NOT NULL";
         if (column.isUnique === true)
             c += " UNIQUE";
-        if (column.isGenerated === true) // don't use skipPrimary here since updates can update already exist primary without auto inc.
+        if (column.isGenerated === true && column.generationStrategy === "increment") // don't use skipPrimary here since updates can update already exist primary without auto inc.
             c += " PRIMARY KEY AUTOINCREMENT";
 
         if (column.default !== undefined && column.default !== null) { // todo: same code in all drivers. make it DRY