+++ test/functional/database-schema/column-types/mysql/entity/Post.ts
@@ -90,9 +90,6 @@ export class Post {
     @Column("json")
     json: Object;
 
-    @Column("uuid")
-    uuid: string;
-
     @Column("simple-array")
     simpleArray: string[];